requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Machine Learning
tensorflow==2.14.0
keras==2.14.0
scikit-learn==1.3.2
numpy==1.24.4
opencv-python==4.8.1.78

# Image Processing
Pillow==10.1.0
imageio==2.31.6

# Data Analysis
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Web Framework
jinja2==3.1.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File Handling
aiofiles==23.2.1

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging
loguru==0.7.2

# Development Tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
pre-commit==3.6.0

# Progress Bars
tqdm==4.66.1

# Utilities
click==8.1.7
rich==13.7.0

# Optional: GPU Support
# tensorflow-gpu==2.14.0  # Uncomment for GPU support

# Deployment
docker==6.1.3

# Monitoring
prometheus-client==0.19.0

# Database (optional)
sqlalchemy==2.0.23
alembic==1.13.1